feat: refactor form state
